fix rep

+++ OpenNos.DAL.EF.MySQL/DB/Character.cs
@@ -35,7 +35,7 @@ namespace OpenNos.DAL.EF.MySQL.DB
         public int Hp { get; set; }
         public int Mp { get; set; }
         public int ArenaWinner { get; set; }
-        public int Reput { get; set; }
+        public long Reput { get; set; }
         public int Dignite { get; set; }
         public long Gold { get; set; }
         public int Backpack { get; set; }
